move util env var
