+++ internal/command/flags.go
@@ -25,8 +25,8 @@ import (
 	"github.com/psiemens/sconfig"
 	"github.com/spf13/cobra"
 
+	"github.com/onflow/flow-cli/internal/util"
 	"github.com/onflow/flow-cli/pkg/flowkit/config"
-	"github.com/onflow/flow-cli/pkg/flowkit/util"
 )
 
 // GlobalFlags contains all global flags definitions.
