+++ internal/util/util.go
@@ -5,6 +5,8 @@ import (
 	"os"
 )
 
+const EnvPrefix = "FLOW"
+
 func Exit(code int, msg string) {
 	fmt.Fprintln(os.Stderr, msg)
 	os.Exit(code)
